+++ R/scrub.R
@@ -50,6 +50,10 @@ scrub.cor_bp <- function(data) {
         dplyr::tbl_df()
 }
 
+scrub.pls_bp <- function(data, ...) {
+    attr(data, 'specs')$results
+}
+
 # Polish ------------------------------------------------------------------
 
 #' Do some final polishing of the scrubbed mason analysis data.
@@ -72,16 +76,16 @@ scrub.cor_bp <- function(data) {
 #' polish_transform_estimates(ds, function(x) exp(x))
 NULL
 
-#' @describeIn polish \code{polish_renaming} simply takes a function, most
-#'   likely one that uses \code{\link[base]{gsub}}, and uses that to search and
+#' @describeIn polish `polish_renaming` simply takes a function, most
+#'   likely one that uses [base::gsub()], and uses that to search and
 #'   replace words, etc, in the specified columns.
-#' @param renaming.fun A function, typically with \code{\link[base]{gsub}}, that
+#' @param renaming.fun A function, typically with [base::gsub()], that
 #'   searches and replaces strings.
 #' @param columns The columns to apply the renaming function to. Defaults to
 #'   columns that are a factor or character vectors.
 #' @export
 polish_renaming <- function(data, renaming.fun, columns = NULL) {
-    assertive::assert_is_function(renaming.fun)
+    stopifnot(is.function(renaming.fun))
 
     if (is.null(columns)) {
         columns <-
@@ -92,14 +96,14 @@ polish_renaming <- function(data, renaming.fun, columns = NULL) {
             which() %>%
             names()
     } else {
-        assertive::assert_is_character(columns)
+        stopifnot(is.character(columns))
     }
 
-    dplyr::mutate_each_(data, dplyr::funs(renaming.fun), columns)
+    dplyr::mutate_at(data, dplyr::vars(columns), dplyr::funs(renaming.fun))
 }
 
-#' @describeIn polish \code{polish_filter} is basically a thin wrapper around
-#' \code{\link[dplyr]{filter}}, but using \code{\link[base]{grepl}} for the
+#' @describeIn polish `polish_filter` is basically a thin wrapper around
+#' [dplyr::filter()], but using [base::grepl()] for the
 #' pattern searching.
 #'
 #' @param keep.pattern Rows to keep based on a regular expression pattern.
@@ -109,25 +113,25 @@ polish_filter <- function(data, keep.pattern, column) {
     dplyr::filter(data, grepl(keep.pattern, data[[column]], ignore.case = TRUE))
 }
 
-#' @describeIn polish \code{polish_transform_estimates} is simply a thin wrapper
-#'   around \code{\link[dplyr]{mutate_each}}.
+#' @describeIn polish `polish_transform_estimates` is simply a thin wrapper
+#'   around [dplyr::mutate_at()].
 #'
 #' @param transform.fun A function to modify continuous variable columns.
 #' @export
 polish_transform_estimates <- function(data, transform.fun) {
-    assertive::assert_is_function(transform.fun)
-    dplyr::mutate_each(
+    stopifnot(is.function(transform.fun))
+    dplyr::mutate_at(
         data,
-        dplyr::funs(transform.fun),
-        dplyr::matches('estimate|std\\.error|conf\\.low|conf\\.high')
+        dplyr::vars(dplyr::matches('estimate|std\\.error|conf\\.low|conf\\.high')),
+        dplyr::funs(transform.fun)
     )
 }
 
-#' @describeIn polish \code{polish_adjust_pvalue} is a thin wrapper around
-#'   \code{\link[dplyr]{mutate}} and \code{\link[stats]{p.adjust}}
+#' @describeIn polish `polish_adjust_pvalue` is a thin wrapper around
+#'   [dplyr::mutate()] and [stats::p.adjust()]
 #'
 #' @param method Correction method for the p-value adjustment
-#'   (\code{\link[stats]{p.adjust}}).
+#'   ([stats::p.adjust()]).
 #' @export
 polish_adjust_pvalue <- function(data, method = 'BH') {
     mutate_tool <- lazyeval::interp('stats::p.adjust(p.value, method)',