+++ NEWS.md
@@ -1,3 +1,11 @@
+# mason 0.2.6
+
+## Bug fixes and other changes
+
+- Fixed changes to new `mutate` functions from dplyr version update
+- Removed assertive package dependency
+- Fixed broom package changes
+
 # mason 0.2.5
 
 ## Hotfixes